requirements.txt
Django==4.2.20
djangorestframework
django-cors-headers
python-dotenv
clerk-backend-api